requirements.txt
langchain
pypdf 
chromadb 
sentence-transformers 
python-decouple 
tiktoken 
lark